Refactor and add documentation.
