+++ analysis_options.yaml
@@ -1,0 +1,3 @@
+linter:
+  rules:
+    - public_member_api_docs
