+++ lib/number_inc_dec.dart
@@ -1,1 +1,2 @@
 export 'src/number_increment_decrement.dart';
+export 'src/prefabbed_widgets.dart';
