+++ lib/src/number_increment_decrement.dart
@@ -2,7 +2,6 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-// import 'package:number_inc_dec/number_inc_dec.dart';
 
 /// A definition for the callback function.
 /// This is the expected function definition for the callbacks after increment and decrement.
@@ -13,453 +12,6 @@ typedef void DiffIncDecCallBack(num newValue);
 /// This is the expected function definition for the callbacks via onSubmitted.
 /// The callback will be passed the new value.
 typedef void ValueCallBack(num newValue);
-
-/// A set of pre-fabricated widgets based on [NumberInputWithIncrementDecrement].
-class NumberInputPrefabbed extends NumberInputWithIncrementDecrement {
-  NumberInputPrefabbed.squaredButtons({
-    @required TextEditingController controller,
-    Key key,
-    bool enabled = true,
-    ButtonArrangement buttonArrangement = ButtonArrangement.rightEnd,
-    num min = 0,
-    num max = double.infinity,
-    num initialValue = 0,
-    num incDecFactor = 1,
-    bool isInt = true,
-    bool autovalidate = true,
-    bool enableMinMaxClamping = false,
-    InputDecoration numberFieldDecoration,
-    Decoration widgetContainerDecoration = const BoxDecoration(),
-    FormFieldValidator<String> validator,
-    TextStyle style,
-    IconData incIcon = Icons.arrow_drop_up,
-    IconData decIcon = Icons.arrow_drop_down,
-    int fractionDigits = 2,
-    double scaleWidth = 1.0,
-    double scaleHeight = 1.0,
-    double incIconSize = 28,
-    double decIconSize = 28,
-    Color decIconColor,
-    Color incIconColor,
-    DiffIncDecCallBack onDecrement,
-    DiffIncDecCallBack onIncrement,
-    ValueCallBack onSubmitted,
-    ValueCallBack onChanged,
-    bool separateIcons = false,
-    Color incDecBgColor = Colors.lightGreen,
-  }) : super(
-          controller: controller,
-          key: key,
-          enabled: enabled,
-          enableMinMaxClamping: enableMinMaxClamping,
-          buttonArrangement: buttonArrangement,
-          numberFieldDecoration: numberFieldDecoration,
-          widgetContainerDecoration: widgetContainerDecoration,
-          validator: validator,
-          style: style,
-          incIcon: incIcon,
-          decIcon: decIcon,
-          fractionDigits: fractionDigits,
-          scaleHeight: scaleHeight,
-          scaleWidth: scaleWidth,
-          incIconSize: incIconSize,
-          decIconSize: decIconSize,
-          decIconColor: decIconColor,
-          incIconColor: incIconColor,
-          onDecrement: onDecrement,
-          onIncrement: onIncrement,
-          onSubmitted: onSubmitted,
-          onChanged: onChanged,
-          separateIcons: separateIcons,
-          incDecBgColor: incDecBgColor,
-          incIconDecoration: BoxDecoration(
-            color: incDecBgColor,
-            border: Border.all(
-              color: Colors.black,
-              width: 0.5,
-            ),
-          ),
-          decIconDecoration: BoxDecoration(
-            color: incDecBgColor,
-            border: Border.all(
-              color: Colors.black,
-              width: 0.5,
-            ),
-          ),
-        );
-
-  NumberInputPrefabbed.leafyButtons({
-    @required TextEditingController controller,
-    Key key,
-    bool enabled = true,
-    ButtonArrangement buttonArrangement = ButtonArrangement.rightEnd,
-    num min = 0,
-    num max = double.infinity,
-    num initialValue = 0,
-    num incDecFactor = 1,
-    bool isInt = true,
-    bool autovalidate = true,
-    bool enableMinMaxClamping = false,
-    InputDecoration numberFieldDecoration,
-    Decoration widgetContainerDecoration = const BoxDecoration(),
-    FormFieldValidator<String> validator,
-    TextStyle style,
-    IconData incIcon = Icons.arrow_drop_up,
-    IconData decIcon = Icons.arrow_drop_down,
-    int fractionDigits = 2,
-    double scaleWidth = 1.0,
-    double scaleHeight = 1.0,
-    double incIconSize = 28,
-    double decIconSize = 28,
-    Color decIconColor,
-    Color incIconColor,
-    DiffIncDecCallBack onDecrement,
-    DiffIncDecCallBack onIncrement,
-    ValueCallBack onSubmitted,
-    ValueCallBack onChanged,
-    bool separateIcons = true,
-    Color incDecBgColor = Colors.lightGreen,
-  }) : super(
-          controller: controller,
-          key: key,
-          enabled: enabled,
-          enableMinMaxClamping: enableMinMaxClamping,
-          buttonArrangement: buttonArrangement,
-          numberFieldDecoration: numberFieldDecoration,
-          widgetContainerDecoration: widgetContainerDecoration,
-          validator: validator,
-          style: style,
-          incIcon: incIcon,
-          decIcon: decIcon,
-          fractionDigits: fractionDigits,
-          scaleHeight: scaleHeight,
-          scaleWidth: scaleWidth,
-          incIconSize: incIconSize,
-          decIconSize: decIconSize,
-          decIconColor: decIconColor,
-          incIconColor: incIconColor,
-          onDecrement: onDecrement,
-          onIncrement: onIncrement,
-          onSubmitted: onSubmitted,
-          onChanged: onChanged,
-          separateIcons: separateIcons,
-          incDecBgColor: incDecBgColor,
-          incIconDecoration: BoxDecoration(
-            color: incDecBgColor,
-            borderRadius: buttonArrangement == ButtonArrangement.rightEnd
-                ? BorderRadius.only(
-                    bottomRight: Radius.circular(10),
-                    topLeft: Radius.circular(10),
-                  )
-                : BorderRadius.only(
-                    topRight: Radius.circular(10),
-                    bottomLeft: Radius.circular(10),
-                  ),
-            border: Border.all(
-              color: Colors.black,
-              width: 0.25,
-            ),
-          ),
-          decIconDecoration: BoxDecoration(
-            color: incDecBgColor,
-            borderRadius: buttonArrangement == ButtonArrangement.rightEnd
-                ? BorderRadius.only(
-                    topRight: Radius.circular(10),
-                    bottomLeft: Radius.circular(10),
-                  )
-                : BorderRadius.only(
-                    bottomRight: Radius.circular(10),
-                    topLeft: Radius.circular(10),
-                  ),
-            border: Border.all(
-              color: Colors.black,
-              width: 0.25,
-            ),
-          ),
-        );
-
-  NumberInputPrefabbed.directionalButtons({
-    @required TextEditingController controller,
-    Key key,
-    bool enabled = true,
-    ButtonArrangement buttonArrangement = ButtonArrangement.rightEnd,
-    num min = 0,
-    num max = double.infinity,
-    num initialValue = 0,
-    num incDecFactor = 1,
-    bool isInt = true,
-    bool autovalidate = true,
-    bool enableMinMaxClamping = false,
-    InputDecoration numberFieldDecoration,
-    Decoration widgetContainerDecoration = const BoxDecoration(),
-    FormFieldValidator<String> validator,
-    TextStyle style,
-    IconData incIcon = Icons.arrow_drop_up,
-    IconData decIcon = Icons.arrow_drop_down,
-    int fractionDigits = 2,
-    double scaleWidth = 1.0,
-    double scaleHeight = 1.0,
-    double incIconSize = 28,
-    double decIconSize = 28,
-    Color decIconColor,
-    Color incIconColor,
-    DiffIncDecCallBack onDecrement,
-    DiffIncDecCallBack onIncrement,
-    ValueCallBack onSubmitted,
-    ValueCallBack onChanged,
-    bool separateIcons = true,
-    Color incDecBgColor = Colors.lightGreen,
-  }) : super(
-          controller: controller,
-          key: key,
-          enabled: enabled,
-          enableMinMaxClamping: enableMinMaxClamping,
-          buttonArrangement: buttonArrangement,
-          numberFieldDecoration: numberFieldDecoration,
-          widgetContainerDecoration: widgetContainerDecoration,
-          validator: validator,
-          style: style,
-          incIcon: incIcon,
-          decIcon: decIcon,
-          fractionDigits: fractionDigits,
-          scaleHeight: scaleHeight,
-          scaleWidth: scaleWidth,
-          incIconSize: incIconSize,
-          decIconSize: decIconSize,
-          decIconColor: decIconColor,
-          incIconColor: incIconColor,
-          onDecrement: onDecrement,
-          onIncrement: onIncrement,
-          onSubmitted: onSubmitted,
-          onChanged: onChanged,
-          separateIcons: separateIcons,
-          incDecBgColor: incDecBgColor,
-          incIconDecoration: BoxDecoration(
-            color: incDecBgColor,
-            borderRadius: BorderRadius.only(
-              topRight: Radius.circular(15),
-              topLeft: Radius.circular(15),
-            ),
-            border: Border.all(
-              color: Colors.black,
-              width: 0.25,
-            ),
-          ),
-          decIconDecoration: BoxDecoration(
-            color: incDecBgColor,
-            borderRadius: BorderRadius.only(
-              bottomRight: Radius.circular(15),
-              bottomLeft: Radius.circular(15),
-            ),
-            border: Border.all(
-              color: Colors.black,
-              width: 0.25,
-            ),
-          ),
-        );
-
-  NumberInputPrefabbed.roundedEdgeButtons({
-    @required TextEditingController controller,
-    Key key,
-    bool enabled = true,
-    ButtonArrangement buttonArrangement = ButtonArrangement.rightEnd,
-    num min = 0,
-    num max = double.infinity,
-    num initialValue = 0,
-    num incDecFactor = 1,
-    bool isInt = true,
-    bool autovalidate = true,
-    bool enableMinMaxClamping = false,
-    InputDecoration numberFieldDecoration,
-    Decoration widgetContainerDecoration = const BoxDecoration(),
-    FormFieldValidator<String> validator,
-    TextStyle style,
-    IconData incIcon = Icons.arrow_drop_up,
-    IconData decIcon = Icons.arrow_drop_down,
-    int fractionDigits = 2,
-    double scaleWidth = 1.0,
-    double scaleHeight = 1.0,
-    double incIconSize = 27,
-    double decIconSize = 27,
-    Color decIconColor,
-    Color incIconColor,
-    DiffIncDecCallBack onDecrement,
-    DiffIncDecCallBack onIncrement,
-    ValueCallBack onSubmitted,
-    ValueCallBack onChanged,
-    bool separateIcons = true,
-    Color incDecBgColor = Colors.lightGreen,
-  }) : super(
-          controller: controller,
-          key: key,
-          enabled: enabled,
-          min: min,
-          max: max,
-          initialValue: initialValue,
-          incDecFactor: incDecFactor,
-          isInt: isInt,
-          autovalidate: autovalidate,
-          enableMinMaxClamping: enableMinMaxClamping,
-          buttonArrangement: buttonArrangement,
-          widgetContainerDecoration: widgetContainerDecoration,
-          validator: validator,
-          style: style,
-          incIcon: incIcon,
-          decIcon: decIcon,
-          fractionDigits: fractionDigits,
-          scaleHeight: scaleHeight,
-          scaleWidth: scaleWidth,
-          incIconSize: incIconSize,
-          decIconSize: decIconSize,
-          decIconColor: decIconColor,
-          incIconColor: incIconColor,
-          onDecrement: onDecrement,
-          onIncrement: onIncrement,
-          onSubmitted: onSubmitted,
-          onChanged: onChanged,
-          separateIcons: separateIcons,
-          incDecBgColor: incDecBgColor,
-          numberFieldDecoration:
-              buttonArrangement == ButtonArrangement.rightEnd ||
-                      buttonArrangement == ButtonArrangement.leftEnd
-                  ? InputDecoration(
-                      border: OutlineInputBorder(
-                        borderRadius:
-                            buttonArrangement == ButtonArrangement.rightEnd
-                                ? BorderRadius.only(
-                                    topLeft: Radius.circular(5),
-                                    bottomLeft: Radius.circular(5),
-                                  )
-                                : BorderRadius.only(
-                                    topRight: Radius.circular(5),
-                                    bottomRight: Radius.circular(5),
-                                  ),
-                      ),
-                    )
-                  : null,
-          incIconDecoration: BoxDecoration(
-            color: incDecBgColor,
-            borderRadius: buttonArrangement == ButtonArrangement.rightEnd
-                ? BorderRadius.only(
-                    topRight: Radius.circular(5),
-                  )
-                : buttonArrangement == ButtonArrangement.incLeftDecRight
-                    ? BorderRadius.only(
-                        topLeft: Radius.circular(5),
-                        bottomLeft: Radius.circular(5),
-                      )
-                    : buttonArrangement == ButtonArrangement.incRightDecLeft
-                        ? BorderRadius.only(
-                            topRight: Radius.circular(5),
-                            bottomRight: Radius.circular(5),
-                          )
-                        : BorderRadius.only(
-                            topLeft: Radius.circular(5),
-                          ),
-            border: Border.all(
-              color: Colors.black,
-              width: 0.25,
-            ),
-          ),
-          decIconDecoration: BoxDecoration(
-            color: incDecBgColor,
-            borderRadius: buttonArrangement == ButtonArrangement.rightEnd
-                ? BorderRadius.only(
-                    bottomRight: Radius.circular(5),
-                  )
-                : buttonArrangement == ButtonArrangement.incLeftDecRight
-                    ? BorderRadius.only(
-                        topRight: Radius.circular(5),
-                        bottomRight: Radius.circular(5),
-                      )
-                    : buttonArrangement == ButtonArrangement.incRightDecLeft
-                        ? BorderRadius.only(
-                            topLeft: Radius.circular(5),
-                            bottomLeft: Radius.circular(5),
-                          )
-                        : BorderRadius.only(
-                            bottomLeft: Radius.circular(5),
-                          ),
-            border: Border.all(
-              color: Colors.black,
-              width: 0.25,
-            ),
-          ),
-        );
-  NumberInputPrefabbed.roundedButtons({
-    @required TextEditingController controller,
-    Key key,
-    bool enabled = true,
-    ButtonArrangement buttonArrangement = ButtonArrangement.rightEnd,
-    num min = 0,
-    num max = double.infinity,
-    num initialValue = 0,
-    num incDecFactor = 1,
-    bool isInt = true,
-    bool autovalidate = true,
-    bool enableMinMaxClamping = false,
-    InputDecoration numberFieldDecoration,
-    Decoration widgetContainerDecoration = const BoxDecoration(),
-    FormFieldValidator<String> validator,
-    TextStyle style,
-    IconData incIcon = Icons.arrow_drop_up,
-    IconData decIcon = Icons.arrow_drop_down,
-    int fractionDigits = 2,
-    double scaleWidth = 1.0,
-    double scaleHeight = 1.0,
-    double incIconSize = 27,
-    double decIconSize = 27,
-    Color decIconColor,
-    Color incIconColor,
-    DiffIncDecCallBack onDecrement,
-    DiffIncDecCallBack onIncrement,
-    ValueCallBack onSubmitted,
-    ValueCallBack onChanged,
-    bool separateIcons = true,
-    Color incDecBgColor = Colors.lightGreen,
-  }) : super(
-          controller: controller,
-          key: key,
-          enabled: enabled,
-          enableMinMaxClamping: enableMinMaxClamping,
-          buttonArrangement: buttonArrangement,
-          numberFieldDecoration: numberFieldDecoration,
-          widgetContainerDecoration: widgetContainerDecoration,
-          validator: validator,
-          style: style,
-          incIcon: incIcon,
-          decIcon: decIcon,
-          fractionDigits: fractionDigits,
-          scaleHeight: scaleHeight,
-          scaleWidth: scaleWidth,
-          incIconSize: incIconSize,
-          decIconSize: decIconSize,
-          decIconColor: decIconColor,
-          incIconColor: incIconColor,
-          onDecrement: onDecrement,
-          onIncrement: onIncrement,
-          onSubmitted: onSubmitted,
-          onChanged: onChanged,
-          separateIcons: separateIcons,
-          incDecBgColor: incDecBgColor,
-          incIconDecoration: BoxDecoration(
-            color: incDecBgColor,
-            borderRadius: BorderRadius.all(Radius.circular(50)),
-            border: Border.all(
-              color: Colors.black,
-              width: 0.25,
-            ),
-          ),
-          decIconDecoration: BoxDecoration(
-            color: incDecBgColor,
-            borderRadius: BorderRadius.all(Radius.circular(50)),
-            border: Border.all(
-              color: Colors.black,
-              width: 0.25,
-            ),
-          ),
-        );
-}
 
 /// A set of button arrangements for the widget.
 enum ButtonArrangement {
@@ -476,10 +28,17 @@ enum ButtonArrangement {
   incRightDecLeft,
 }
 
+/// {@template core_widget_doc}
 /// A simple widget that wraps a TextFormField, exposes it as a number field and
 /// adds buttons to increment and decrement the numeric value. Performs some
 /// basic validation, formatting to [fractionDigits] and clamping the values
 /// between [min] and [max] when enabled through [enableMinMaxClamping].
+///
+/// e.g. Default widget appearance.
+///
+/// ![NumberInputWithIncrementDecrement](https://raw.githubusercontent.com/Abhilash-Chandran/number_inc_dec/doc-assets/gifs/example1.gif)
+///
+/// {@endtemplate}
 
 class NumberInputWithIncrementDecrement extends StatefulWidget {
   /// Key to be used for this widget.
@@ -623,6 +182,8 @@ class NumberInputWithIncrementDecrement extends StatefulWidget {
 
   /// Background color of increment decrement buttons.
   final Color incDecBgColor;
+
+  /// {@macro core_widget_doc}
   NumberInputWithIncrementDecrement({
     @required this.controller,
     this.key,
